requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic

#RAG dependencies:
groq
chromadb
sentence-transformers
